refactor: Separate CallvanNotificationUiItem into model package
Move CallvanNotificationUiItem data class from component to
ui/notification/model to follow project conventions and fix
detekt MatchingDeclarationName violation.
