+++ feature/callvan/src/main/java/in/koreatech/koin/feature/callvan/ui/notification/model/CallvanNotificationUiItem.kt
@@ -1,0 +1,9 @@
+package `in`.koreatech.koin.feature.callvan.ui.notification.model
+
+data class CallvanNotificationUiItem(
+    val id: Int,
+    val title: String,
+    val routeInfo: String,
+    val message: String,
+    val isRead: Boolean
+)
